Release v1.25
- Adding skipping clone table for drop targets in MySQL and SQL Server

+++ plugins/plugins/MySQLSubject.java
@@ -186,7 +186,8 @@ public class MySQLSubject extends ExperimentSubject {
 			for (int i = 0; i < vecTables.size(); i++) {
 				String tblName = (String) vecTables.get(i);
 				Main._logger.outputLog("installed tableName: " + tblName);
-				dropTable(tblName);
+				if(!tblName.contains(Constants.CLONE_TABLE_PREFIX))
+					dropTable(tblName);
 			}
 		} catch (SQLException e) {
 			e.printStackTrace();
